changed all loggers from instance variables to class/module constants
since none of the loggers need per-instance configurations and their instantiation is basically only there to differentiate between the different class-types or modules that create a log entry, there is no need to create per-instance variables. as an additional bonus it is also much better visibly to drop the @'s.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>